Day 09 - refactor
